+++ src/main/java/nextstep/subway/line/application/exceptions/AlreadyExistsLineNameException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.application;
+package nextstep.subway.line.application.exceptions;
 
 public class AlreadyExistsLineNameException extends RuntimeException {
 	public AlreadyExistsLineNameException(String message) {
