+++ src/main/java/nextstep/subway/line/application/exceptions/NotFoundLineException.java
@@ -1,0 +1,5 @@
+package nextstep.subway.line.application.exceptions;
+
+public class NotFoundLineException extends RuntimeException {
+
+}
