+++ src/main/java/nextstep/subway/line/ui/LineController.java
@@ -1,6 +1,6 @@
 package nextstep.subway.line.ui;
 
-import nextstep.subway.line.application.AlreadyExistsLineNameException;
+import nextstep.subway.line.application.exceptions.AlreadyExistsLineNameException;
 import nextstep.subway.line.application.LineService;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.dto.LineRequest;
